refactor(guid): convert files to typescript
