--- src/__internal__/utils/helpers/guid/package.json
@@ -1,3 +1,0 @@
-{
-  "main": "./guid.js"
-}
